requirements.txt
playwright
beautifulsoup4
lxml
python-slugify
pytest